Using separate react somponent for menu.
