+++ packages/@tinacms/fields/src/Wysiwyg/state/index.ts
@@ -27,9 +27,8 @@ import { tableEditing } from 'prosemirror-tables'
 
 import { inputRules } from './plugins/input-rules'
 import { Translator } from '../Translator'
-import { menu } from './plugins/Menu'
 import { buildKeymap } from './buildKeymap'
-import { links } from './plugins/links'
+// import { links } from './plugins/links'
 import { tablePlugin } from './plugins/Table'
 import { imagePlugin } from './plugins/Image'
 
@@ -37,9 +36,8 @@ export function createEditorState(
   schema: Schema,
   translator: Translator,
   plugins: Plugin[],
-  value: string,
-  theme: any, // TODO: update type
-  sticky?: boolean
+  value: string
+  // theme: any // TODO: update type
 ) {
   return EditorState.create({
     schema,
@@ -48,10 +46,9 @@ export function createEditorState(
       inputRules(schema),
       keymap(buildKeymap(schema, plugins)),
       history(),
-      links(schema, theme),
+      // links(schema, theme),
       dropCursor({ width: 2, color: 'rgb(33, 224, 158)' }),
       gapCursor(),
-      menu(translator, false, theme, sticky),
       tableEditing(),
       tablePlugin,
       imagePlugin,
