+++ packages/@tinacms/fields/src/Wysiwyg/state/plugins/inline-menu/blockControl.tsx
@@ -30,19 +30,22 @@ interface BlockTool {
 export function blockTool(options: BlockTool) {
   const { children, command, typeName, attrs } = options
   return class extends React.Component<
-    { view: EditorView; onClick(): void },
+    { editorView: { view: EditorView }; onClick(): void },
     any
   > {
-    canDo = () => command(this.props.view.state)
+    canDo = () => command(this.props.editorView.view.state)
     onClick = () => {
-      command(this.props.view.state, this.props.view.dispatch)
-      this.props.view.focus()
+      command(
+        this.props.editorView.view.state,
+        this.props.editorView.view.dispatch
+      )
+      this.props.editorView.view.focus()
       this.props.onClick()
     }
     get active(): boolean {
       if (!typeName) return false
 
-      const { state } = this.props.view
+      const { state } = this.props.editorView.view
       const $from = state.selection.$from
       const node = $from.node($from.depth)
       const correctNodeType = node.type.name === typeName
